refactor(S-3.01): extract handle_refresh to auth/refresh.rs (BC-7.4.015)
Move REFRESH_HELP_LINE, refresh_success_payload, RefreshArgs, and
refresh_credentials out of mod.rs. refresh.rs calls into login_token /
login_oauth via super:: imports — no behavioral change. RefreshArgs and
refresh_credentials are pub; refresh_success_payload re-exported under
#[cfg(test)] for regression-pin tests. BC-7.4.015 preserved.
612 unit tests pass.
